refactor: move the handler functions to a separate directory; change directory for files create/edit-car.ts
